app/requirements.txt
streamlit
pandas
numpy
pillow
plotly
scikit-learn
viyapy